refactor: move dev-only code behind build tags
Move CreateTestJob, testOutputGenerators, and registerTestPattern into
dev-tagged files so they are excluded from production builds entirely.

Add devRouteRegistrars extension point in server.go so dev-tagged init()
functions can register additional routes without stubs.
