+++ server/controllers/jobs_controller.go
@@ -22,21 +22,8 @@ import (
 	tally "github.com/uber-go/tally/v4"
 )
 
-// testOutputGenerators holds registered test output generators, keyed by pattern name.
-// Populated via init() in jobs_test_patterns.go (dev builds only).
-var testOutputGenerators map[string]func(chan<- string)
-
 // sseConnectionCount tracks active SSE connections
 var sseConnectionCount atomic.Int32
-
-// registerTestPattern registers a test output generator for dev mode.
-// Called from init() in build-tagged files.
-func registerTestPattern(name string, gen func(chan<- string)) {
-	if testOutputGenerators == nil {
-		testOutputGenerators = make(map[string]func(chan<- string))
-	}
-	testOutputGenerators[name] = gen
-}
 
 type JobIDKeyGenerator struct{}
 
@@ -131,22 +118,22 @@ func (j *JobsController) getProjectJobs(w http.ResponseWriter, r *http.Request) 
 					ActiveNav:       "jobs",
 					ApplyLockActive: applyLockActive,
 				},
-				JobID:        jobID,
-				JobStep:      dbOutput.CommandName,
-				RepoFullName: dbOutput.RepoFullName,
-				RepoOwner:    repoOwner,
-				RepoName:     repoName,
-				PullNum:      dbOutput.PullNum,
-				ProjectPath:  dbOutput.Path,
-				Workspace:    dbOutput.Workspace,
-				Status:       status,
-				Output:       dbOutput.Output,
-				TriggeredBy:  dbOutput.TriggeredBy,
-				BadgeText:    badgeText,
-				BadgeStyle:   badgeStyle,
-				AddCount:     dbOutput.ResourceStats.Add,
-				ChangeCount:  dbOutput.ResourceStats.Change,
-				DestroyCount: dbOutput.ResourceStats.Destroy,
+				JobID:          jobID,
+				JobStep:        dbOutput.CommandName,
+				RepoFullName:   dbOutput.RepoFullName,
+				RepoOwner:      repoOwner,
+				RepoName:       repoName,
+				PullNum:        dbOutput.PullNum,
+				ProjectPath:    dbOutput.Path,
+				Workspace:      dbOutput.Workspace,
+				Status:         status,
+				Output:         dbOutput.Output,
+				TriggeredBy:    dbOutput.TriggeredBy,
+				BadgeText:      badgeText,
+				BadgeStyle:     badgeStyle,
+				AddCount:       dbOutput.ResourceStats.Add,
+				ChangeCount:    dbOutput.ResourceStats.Change,
+				DestroyCount:   dbOutput.ResourceStats.Destroy,
 				PolicyPassed:   dbOutput.PolicyPassed,
 				HasPolicyCheck: dbOutput.CommandName == "policy_check" || dbOutput.PolicyOutput != "",
 			}
@@ -452,111 +439,4 @@ func (j *JobsController) GetProjectJobsSSE(w http.ResponseWriter, r *http.Reques
 		j.Logger.Err(err.Error())
 		errorCounter.Inc(1)
 	}
-}
-
-// CreateTestJob creates a test job for development/testing purposes.
-// Only available in dev mode. Test output generators are registered via
-// init() in jobs_test_patterns.go, which is only compiled with the "dev"
-// build tag. Without the tag, testOutputGenerators is nil and this
-// handler returns 404.
-// Query params:
-//   - pattern: output pattern (default, slow, burst, colors, error, long)
-//   - repo: repository full name (e.g., "acme/infrastructure")
-//   - pr: pull request number
-//   - project: project path
-//   - step: job step (plan, apply)
-func (j *JobsController) CreateTestJob(w http.ResponseWriter, r *http.Request) {
-	if testOutputGenerators == nil {
-		http.Error(w, "Not available", http.StatusNotFound)
-		return
-	}
-
-	patternName := r.URL.Query().Get("pattern")
-	if patternName == "" {
-		patternName = "default"
-	}
-
-	gen, ok := testOutputGenerators[patternName]
-	if !ok {
-		gen, ok = testOutputGenerators["default"]
-		if !ok {
-			http.Error(w, "No test patterns registered", http.StatusNotFound)
-			return
-		}
-	}
-
-	// Parse PR params with defaults
-	repoFullName := r.URL.Query().Get("repo")
-	if repoFullName == "" {
-		repoFullName = "test/repository"
-	}
-
-	prNumStr := r.URL.Query().Get("pr")
-	prNum := 0
-	if prNumStr != "" {
-		if n, err := strconv.Atoi(prNumStr); err == nil {
-			prNum = n
-		}
-	}
-
-	projectPath := r.URL.Query().Get("project")
-	if projectPath == "" {
-		projectPath = "terraform/test"
-	}
-
-	jobStep := r.URL.Query().Get("step")
-	if jobStep == "" {
-		jobStep = "plan"
-	}
-
-	// Extract repo name from full name
-	repoName := repoFullName
-	if parts := strings.Split(repoFullName, "/"); len(parts) == 2 {
-		repoName = parts[1]
-	}
-
-	// Generate unique job ID
-	jobID := fmt.Sprintf("test-%d", time.Now().UnixNano())
-
-	// Build PullInfo for job mapping
-	pullInfo := jobs.PullInfo{
-		PullNum:      prNum,
-		Repo:         repoName,
-		RepoFullName: repoFullName,
-		ProjectName:  "",
-		Path:         projectPath,
-		Workspace:    "default",
-	}
-
-	handler, ok := j.OutputHandler.(*jobs.AsyncProjectCommandOutputHandler)
-	if !ok {
-		j.respond(w, logging.Warn, http.StatusInternalServerError, "Dev test jobs not supported with this output handler")
-		return
-	}
-
-	// Register the job with PR association
-	handler.RegisterTestJob(jobID, pullInfo, jobStep)
-
-	// Create output channel
-	outputChan := make(chan string, 1000)
-
-	// Start generating output in background
-	go func() {
-		// Small delay to allow redirect to complete
-		time.Sleep(100 * time.Millisecond)
-		gen(outputChan)
-		close(outputChan)
-	}()
-
-	// Feed output to handler
-	go func() {
-		for line := range outputChan {
-			handler.SendTestLine(jobID, line)
-		}
-		handler.MarkComplete(jobID)
-	}()
-
-	// Redirect to job page
-	redirectURL := fmt.Sprintf("%s/jobs/%s", j.AtlantisURL.Path, jobID)
-	http.Redirect(w, r, redirectURL, http.StatusFound)
 }
