+++ server/server_dev.go
@@ -8,7 +8,7 @@ import (
 	web_templates "github.com/runatlantis/atlantis/server/controllers/web_templates"
 )
 
-//nolint:gochecknoinits // Dev-only: enables template hot reloading when built with -tags dev.
+//nolint:gochecknoinits // Dev-only: enables template hot reloading and dev routes when built with -tags dev.
 func init() {
 	templatesDir := "server/controllers/web_templates/templates"
 	if _, err := os.Stat(templatesDir); os.IsNotExist(err) {
@@ -17,4 +17,8 @@ func init() {
 	if _, err := os.Stat(templatesDir); err == nil {
 		web_templates.SetDevMode(true, templatesDir)
 	}
+
+	devRouteRegistrars = append(devRouteRegistrars, func(s *Server) {
+		s.Router.HandleFunc("/jobs/test", s.JobsController.CreateTestJob).Methods("GET")
+	})
 }
